requirements.txt
requests>=1.1.0
nose>=1.3.0
nose-testconfig>=0.9
